relocate IrisLootEvent to the proper package
